backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
openai==1.54.4
httpx==0.27.2
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.10.1
Pillow==11.0.0
